finished breaking it all out
